+++ src/index.ts
@@ -56,7 +56,7 @@ export default class DriveBridge extends Plugin {
 
 	async onload() {
 		const settings: Settings = {
-			asymmetricStorage: true,
+			asymmetricStorage: false,
 			avoidAutoSyncWhenOffline: true,
 			confirmDeleteInAutoSync: true,
 			confirmTasksInSync: true,